requirements.txt
fastapi==0.111.0
starlette==0.37.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
typing-extensions==4.11.0
google-generativeai==0.8.4
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.41.2
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
redis[hiredis]==5.2.1
chromadb==0.5.0
requests==2.32.3
httpx==0.27.0
anyio==4.3.0
sse-starlette==2.1.0
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==11.1.0
slowapi==0.1.9
cryptography==42.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
tenacity==8.3.0
orjson==3.10.7
numpy==1.26.4
pytest==8.2.0
pytest-asyncio==0.23.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic>=1.13.0